+++ demos/supabase-todolist/lib/widgets/todo_item_widget.dart
@@ -23,7 +23,7 @@ class TodoItemWidget extends StatelessWidget {
 
   Future<void> deleteTodo(TodoItem todo) async {
     if (todo.photoId != null) {
-      attachmentQueue.deletePhoto(todo.photoId!);
+      attachmentQueue.deleteFile(todo.photoId!);
     }
     await todo.delete();
   }