+++ packages/powersync_attachments_helper/lib/src/attachments_queue.dart
@@ -33,6 +33,14 @@ abstract class AbstractAttachmentQueue {
   /// Return true if you want to ignore attachment
   Future<bool> Function(Attachment attachment, Object exception)? onUploadError;
 
+  /// Interval in minutes to periodically run [syncingService.startPeriodicSync]
+  /// Default is 5 minutes
+  int intervalInMinutes;
+
+  /// Provide the subdirectories located on external storage so that they are created
+  /// when the attachment queue is initialized.
+  List<String>? subdirectories;
+
   AbstractAttachmentQueue(
       {required this.db,
       required this.remoteStorage,
@@ -40,7 +48,8 @@ abstract class AbstractAttachmentQueue {
       this.attachmentsQueueTableName = defaultAttachmentsQueueTableName,
       this.onDownloadError,
       this.onUploadError,
-      performInitialSync = true}) {
+      this.intervalInMinutes = 5,
+      this.subdirectories}) {
     attachmentsService = AttachmentsService(
         db, localStorage, attachmentDirectoryName, attachmentsQueueTableName);
     syncingService = SyncingService(
@@ -52,11 +61,11 @@ abstract class AbstractAttachmentQueue {
   /// Set the file extension if you are using a different file type
   StreamSubscription<void> watchIds({String fileExtension = 'jpg'});
 
-  /// Create a function to save photos using the attachment queue
-  Future<Attachment> savePhoto(String photoId, int size);
+  /// Create a function to save files using the attachment queue
+  Future<Attachment> saveFile(String fileId, int size);
 
-  /// Create a function to delete photos using the attachment queue
-  Future<Attachment> deletePhoto(String photoId);
+  /// Create a function to delete files using the attachment queue
+  Future<Attachment> deleteFile(String fileId);
 
   /// Initialize the attachment queue by
   /// 1. Creating attachments directory
@@ -66,8 +75,16 @@ abstract class AbstractAttachmentQueue {
     // Ensure the directory where attachments are downloaded, exists
     await localStorage.makeDir(await getStorageDirectory());
 
+    if (subdirectories != null) {
+      for (String subdirectory in subdirectories!) {
+        await localStorage
+            .makeDir('${await getStorageDirectory()}/$subdirectory');
+      }
+    }
+
     watchIds();
     syncingService.watchAttachments();
+    syncingService.startPeriodicSync(intervalInMinutes);
 
     db.statusStream.listen((status) {
       if (db.currentStatus.connected) {